requirements-integration.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements-integration.in
#
aiohappyeyeballs==2.3.4
    # via aiohttp
aiohttp==3.10.0
    # via -r requirements-integration.in
aiosignal==1.3.1
    # via aiohttp
anyio==4.4.0
    # via httpx
asttokens==2.4.1
    # via stack-data
async-timeout==4.0.3
    # via aiohttp
attrs==23.2.0
    # via
    #   aiohttp
    #   jsonschema
backcall==0.2.0
    # via ipython
bcrypt==4.2.0
    # via paramiko
cachetools==5.4.0
    # via google-auth
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   kubernetes
    #   requests
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
charmed-kubeflow-chisme==0.4.3
    # via -r requirements-integration.in
charset-normalizer==3.3.2
    # via requests
cryptography==43.0.0
    # via paramiko
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
deepdiff==6.2.1
    # via charmed-kubeflow-chisme
exceptiongroup==1.2.2
    # via
    #   anyio
    #   pytest
executing==2.0.1
    # via stack-data
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
google-auth==2.32.0
    # via kubernetes
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via lightkube
hvac==2.3.0
    # via juju
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-resources==6.4.0
    # via jsonschema
iniconfig==2.0.0
    # via pytest
ipdb==0.13.13
    # via pytest-operator
ipython==8.12.3
    # via ipdb
jedi==0.19.1
    # via ipython
jinja2==3.1.4
    # via
    #   -r requirements-integration.in
    #   charmed-kubeflow-chisme
    #   pytest-operator
jsonschema==4.17.3
    # via serialized-data-interface
juju==3.5.2.0
    # via
    #   -r requirements-integration.in
    #   charmed-kubeflow-chisme
    #   pytest-operator
kubernetes==30.1.0
    # via juju
lightkube==0.15.3
    # via charmed-kubeflow-chisme
lightkube-models==1.30.0.8
    # via lightkube
macaroonbakery==1.3.4
    # via juju
markupsafe==2.1.5
    # via jinja2
matplotlib-inline==0.1.7
    # via ipython
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
oauthlib==3.2.2
    # via
    #   kubernetes
    #   requests-oauthlib
ops==2.15.0
    # via
    #   charmed-kubeflow-chisme
    #   serialized-data-interface
ordered-set==4.1.0
    # via deepdiff
packaging==24.1
    # via
    #   juju
    #   pytest
paramiko==3.4.0
    # via juju
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pkgutil-resolve-name==1.3.10
    # via jsonschema
pluggy==1.5.0
    # via pytest
prompt-toolkit==3.0.47
    # via ipython
protobuf==5.27.3
    # via macaroonbakery
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyasn1==0.6.0
    # via
    #   juju
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pycparser==2.22
    # via cffi
pygments==2.18.0
    # via ipython
pymacaroons==0.13.0
    # via macaroonbakery
pynacl==1.5.0
    # via
    #   macaroonbakery
    #   paramiko
    #   pymacaroons
pyrfc3339==1.1
    # via
    #   juju
    #   macaroonbakery
pyrsistent==0.20.0
    # via jsonschema
pytest==8.3.2
    # via
    #   pytest-asyncio
    #   pytest-operator
pytest-asyncio==0.21.2
    # via pytest-operator
pytest-operator==0.35.0
    # via -r requirements-integration.in
python-dateutil==2.9.0.post0
    # via kubernetes
pytz==2024.1
    # via pyrfc3339
pyyaml==6.0.1
    # via
    #   juju
    #   kubernetes
    #   lightkube
    #   ops
    #   pytest-operator
    #   serialized-data-interface
requests==2.32.3
    # via
    #   -r requirements-integration.in
    #   hvac
    #   kubernetes
    #   macaroonbakery
    #   requests-oauthlib
    #   serialized-data-interface
requests-oauthlib==2.0.0
    # via kubernetes
rsa==4.9
    # via google-auth
ruamel-yaml==0.18.6
    # via charmed-kubeflow-chisme
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
serialized-data-interface==0.7.0
    # via charmed-kubeflow-chisme
six==1.16.0
    # via
    #   asttokens
    #   kubernetes
    #   macaroonbakery
    #   pymacaroons
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
stack-data==0.6.3
    # via ipython
tenacity==9.0.0
    # via charmed-kubeflow-chisme
tomli==2.0.1
    # via
    #   ipdb
    #   pytest
toposort==1.10
    # via juju
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.12.2
    # via
    #   anyio
    #   ipython
    #   typing-inspect
typing-inspect==0.9.0
    # via juju
urllib3==2.2.2
    # via
    #   kubernetes
    #   requests
wcwidth==0.2.13
    # via prompt-toolkit
websocket-client==1.8.0
    # via
    #   kubernetes
    #   ops
websockets==12.0
    # via juju
yarl==1.9.4
    # via aiohttp
zipp==3.19.2
    # via importlib-resources
